+++ orwynn/sql/constants.py
@@ -1,0 +1,8 @@
+from pykit.cls import Static
+
+
+class PSQLStatementConstants(Static):
+    RecreatePublicSchemaCascade: str = (
+        "DROP SCHEMA public CASCADE;"
+        " CREATE SCHEMA public;"
+    )